cant get this to run; errors with normal variate
I installed the normal module and it says normal variate isn't defined; I also tried doing this with numpy. I'll work on understanding the bigger picture of what's going on in the code and working on where we left off last time. I'm guessing it's just a simple import error I'm making or something like that